[refactor] Re-implement http adapter.
The broadcaster is dead or more exactly, has been merged into the adapter. The http recipient now holds the contact method as well such that the Send method is completly protocol agnostic. Data are sent as an octet-stream and the interpretation is under the receiver responsabilities. Handlers to listen on given endpoints can be registered and can also generate registrations