Fix component startup when a monitor server is down
Previously, monitor clients were synchronously initialized during
component startup. As a result, if a monitor server was down during
during startup, the component would fail to start. This introduces a
registry that encapsulates the initialization of monitor clients and
exposes a simple API that allows components to retrieve the monitor
clients that are ready at that point in time.